Create hooks for guild & page actions.
